+++ lib/logging_appenders.dart
@@ -1,8 +1,13 @@
+/// Basic logging appenders for darts logging framework.
+///
+/// The simplest use case is using the [PrintAppender.setupLogging]
+/// to create a logger which logs to stdout.
 library logging_appenders;
 
 export 'src/base_appender.dart' show LoggingAppenders, BaseLogAppender;
 export 'src/logrecord_formatter.dart';
-export 'src/print_appender.dart';
-export 'src/remote/logzio_appender.dart';
-export 'src/remote/loki_appender.dart';
-export 'src/rotating_file_appender.dart';
+export 'src/print_appender.dart' show PrintAppender;
+export 'src/remote/logzio_appender.dart' show LogzIoApiAppender;
+export 'src/remote/loki_appender.dart' show LokiApiAppender;
+export 'src/rotating_file_appender.dart'
+    show AsyncInitializingLogHandler, RotatingFileAppender;